refactor: add FormWrapper component, replace login schema in separate file
