--- src/features/authentication/login/model/login.ts
@@ -1,2 +1,0 @@
-const temp = (): void => {};
-export default temp;
