+++ src/features/authentication/login/ui/LoginForm.tsx
@@ -1,22 +1,14 @@
 import { Controller, SubmitHandler, useForm } from 'react-hook-form';
 
-import { Link } from 'react-router-dom';
-
 import { yupResolver } from '@hookform/resolvers/yup';
 
-import { ObjectSchema, object } from 'yup';
-
-import styles from './LoginForm.module.scss';
-
-import emailSchema from '../../../../shared/validation/model/emailSchema';
-import passwordSchema from '../../../../shared/validation/model/passwordSchema';
+import { ObjectSchema } from 'yup';
 
 import { TextInput, PasswordInput, PasswordErrors } from '../../../../shared/ui';
-import Button from '../../../../shared/ui/button/Button';
 
-import RoutesName from '../../../../shared/routing';
+import FormWrapper from '../../../../shared/ui/form/FormWrapper';
+import loginSchema from '../model/loginSchema';
 
-const userSchema = object().shape({ ...emailSchema.fields, ...passwordSchema.fields });
 interface LoginUserFields {
   email: string;
   password: string;
@@ -28,7 +20,7 @@ function LoginForm(): JSX.Element {
     control,
     formState: { errors },
   } = useForm<LoginUserFields>({
-    resolver: yupResolver(userSchema as ObjectSchema<LoginUserFields>),
+    resolver: yupResolver(loginSchema as ObjectSchema<LoginUserFields>),
     mode: 'onChange',
     defaultValues: {
       email: '',
@@ -41,38 +33,30 @@ function LoginForm(): JSX.Element {
   };
 
   return (
-    <form className={styles.formWrapper} onSubmit={handleSubmit(onSubmit)}>
-      <h3 className={styles.title}>Log in to your account</h3>
-      <div className={styles.line} />
-      <Controller
-        name="email"
-        control={control}
-        rules={{ required: true }}
-        render={({ field }): JSX.Element => {
-          return <TextInput id="1" placeholder="Email" label="Email *" {...field} error={errors.email} />;
-        }}
-      />
-      <Controller
-        name="password"
-        control={control}
-        rules={{ required: true }}
-        render={({ field, fieldState }): JSX.Element => {
-          return (
-            <>
-              <PasswordInput id="2" placeholder="Password" label="Password *" {...field} error={errors.password} />
-              {fieldState.isDirty && errors.password && <PasswordErrors value={field.value || ''} />}
-            </>
-          );
-        }}
-      />
-      <p className={styles.signUpText}>
-        Already have an account?&nbsp;
-        <Link to={RoutesName.registration}>Sign Up</Link>
-      </p>
-      <div className={`${styles.line} ${styles.bottom}`} />
-      <Button type="submit" height="48px" width="80%">
-        Sign In
-      </Button>
+    <form onSubmit={handleSubmit(onSubmit)}>
+      <FormWrapper title="Log in to your account">
+        <Controller
+          name="email"
+          control={control}
+          rules={{ required: true }}
+          render={({ field }): JSX.Element => {
+            return <TextInput id="1" placeholder="Email" label="Email *" {...field} error={errors.email} />;
+          }}
+        />
+        <Controller
+          name="password"
+          control={control}
+          rules={{ required: true }}
+          render={({ field, fieldState }): JSX.Element => {
+            return (
+              <>
+                <PasswordInput id="2" placeholder="Password" label="Password *" {...field} error={errors.password} />
+                {fieldState.isDirty && errors.password && <PasswordErrors value={field.value || ''} />}
+              </>
+            );
+          }}
+        />
+      </FormWrapper>
     </form>
   );
 }
