+++ src/shared/ui/form/FormWrapper.module.scss
@@ -1,4 +1,4 @@
-@import "../../../../app/styles/variables.scss";
+@import "../../../app//styles/variables.scss";
 
 .formWrapper {
   display: flex;
