+++ src/widgets/body/ui/LayoutBody.module.scss
@@ -1,4 +1,8 @@
 .mainBody {
   width: 100%;
   height: 100%;
+
+  form {
+    flex: 0 1 700px;
+  }
 }
